# src/configs/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "InsecureRestAPI",
    "version": "1.0",
    "description": "InsecureRestAPI - An insecure NodeJS/Express/MongoDB REST API.",
    "license": {
      "name": "GPLv3",
      "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
    },
    "contact": {
      "name": "Fortify by Opentext",
      "url": "https://www.opentext.com/en-gb/products/application-security"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "InsecureRestAPI - development "
    }
  ],
  "paths": {
...
    "/api/v1/site/status": {
      "get": {
        "tags": [
          "Site"
        ],
        "summary": "Get the site status",
        "description": "Get the site message of the day",
        "operationId": "getSiteStatus",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/site/register-user": {
      "post": {
        "tags": [
          "Site"
        ],
        "summary": "Register a new user",
        "description": "Register a new user with the site",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "409": {
            "description": "User Already Exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/registerUser"
              }
            }
          }
        }
      }
    },
    "/api/v1/site/subscribe-user": {
      "post": {
        "tags": [
          "Site"
        ],
        "summary": "Subscribe a new user",
        "description": "Subscribe a new user to the newsletter",
        "operationId": "subscribeUser",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "409": {
            "description": "User Already Exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/subscribeUser"
              }
            }
          }
        }
      }
    },
    "/api/v1/site/sign-in": {
      "post": {
        "tags": [
          "Site"
        ],
        "summary": "Sign-in",
        "description": "Sign in as an authorised user",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/jwtJson"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/jwtJson"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/signInUser"
              }
            }
          }
        }
      }
    },
    "/api/v1/site/sign-out": {
      "post": {
        "tags": [
          "Site"
        ],
        "summary": "Sign-out",
        "description": "Sign out an authorised user",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/signOutUser"
              }
            }
          }
        }
      }
    },
    "/api/v1/site/refresh-token": {
      "post": {
        "tags": [
          "Site"
        ],
        "summary": "Refresh token",
        "description": "Refresh users JwtAuthHandler access token",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/refreshUser"
              }
            }
          }
        }
      }
    },
    "/api/v1/site/backup-newsletter-db": {
      "post": {
        "tags": [
          "Site"
        ],
        "summary": "Backup the newsletter database",
        "description": "Compress and backup the newsletter database to the specified file",
        "operationId": "backupNewsletterDb",
        "parameters": [
          {
            "name": "file_path",
            "in": "query",
            "description": "The file to backup the database to. Cannot be empty.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "registerUser": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "example": "auth0|6634fe351045dd069fb881af"
          },
          "first_name": {
            "type": "string",
            "example": "User"
          },
          "last_name": {
            "type": "string",
            "example": "One"
          },
          "email": {
            "type": "string",
            "example": "user1@localhost.com"
          },
          "phone_number": {
            "type": "string",
            "example": "1234567890"
          }
        },
        "required": [
          "user_id",
          "first_name",
          "last_name",
          "email",
          "phone_number"
        ],
        "xml": {
          "name": "registerUser"
        }
      },
      "subscribeUser": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string",
            "example": "User"
          },
          "last_name": {
            "type": "string",
            "example": "One"
          },
          "email": {
            "type": "string",
            "example": "user1@localhost.com"
          }
        },
        "required": [
          "first_name",
          "last_name",
          "email"
        ],
        "xml": {
          "name": "subscribeUser"
        }
      },
      "signInUser": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "string"
          },
          "password": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "xml": {
          "name": "signInUser"
        }
      },
      "signOutUser": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "accessToken"
        ],
        "xml": {
          "name": "signOutUser"
        }
      },
      "refreshUser": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "refreshToken"
        ],
        "xml": {
          "name": "refreshUser"
        }
      },
...
      "success": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string",
            "example": "string"
          },
          "timestamp": {
            "type": "string",
            "example": "YYYY-MM-DDTHH:mm:ssZ"
          },
          "data": {
            "type": "object",
            "properties": {}
          }
        },
        "required": [
          "status",
          "message",
          "timestamp",
          "data"
        ],
        "xml": {
          "name": "success"
        }
      },
      "failure": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "failure"
          },
          "message": {
            "type": "string",
            "example": "string"
          },
          "timestamp": {
            "type": "string",
            "example": "YYYY-MM-DDTHH:mm:ssZ"
          },
          "data": {
            "type": "object",
            "properties": {}
          }
        },
        "required": [
          "status",
          "message",
          "timestamp",
          "data"
        ],
        "xml": {
          "name": "failure"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Access Token"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}